fix(index): stop exporting minimumExpectedToolCount from package root
- Move minimumExpectedToolCount and its HEALTH_CHECK/BASELINE_TOOL_NAMES
  constants out of src/index.ts into a new internal src/health-check.ts,
  importing but not re-exporting them from the root module.
- OpenCode's legacy plugin loader iterates every named export of the
  package root and invokes each as a plugin factory with PluginInput.
  Exporting minimumExpectedToolCount meant it was called with
  PluginInput (not string[]) and its numeric return value pushed into
  the hooks array as if it were a Hooks object. The Array.isArray guard
  added in this branch prevents the resulting .filter() crash but does
  not stop the mis-invocation itself.
- Move the corresponding unit tests to src/health-check.test.ts.

Addresses triage feedback on #894:
https://github.com/alvinunreal/oh-my-opencode-slim/issues/894#issuecomment-5146763258
